Response to #111
@KimPatch Conscious that it would be good to have a more real time back and forth with this! Hopefully we can work through it with updates in GitHub.

I have done a rewrite with your comments in mind (and immediately next to the update!). We are trying to retain the sense of Alex being that novice user of speech control. As such I have made this a much more recent diagnosis for Alex and am trying to talk through his journey of changing how he works. This retains the features of the previous persona but the context might make more sense.

Keen for your thoughts on the latest iteration and happy to work through this more.