Fixes #214.
Sometimes the design of these sites, including the process for purchasing memorabilia, is quite busy with distracting advertisements and complicated checkout forms. However, she has found that she can access these sites on her mobile phone where she can switch the reading mode on her browser to remove a lot of the background clutter. This allows her to focus on the task of choosing what she wants to buy and filling out the form fields correctly.