Include video view sample
Note that the videos do not yet work, including the link to the text transcript. The idea is that activating the video link (the image) would open a modal window to show a larger view of the video.

The intention is to show how they could be integrated into the page.